requirements.txt
Django==1.9.7
psycopg2==2.6.1
dj-database-url==0.4.1
gunicorn==19.6.0

Pillow==3.3.0
tinify==1.3.0

factory-boy==2.7.0
  fake-factory==0.7.2
  python-dateutil==2.5.3
